move code responsible for restructuring docs-only stories to the rendering section, so refs get this behavior as well.
